+++ src/css/styles.module.scss
@@ -66,7 +66,7 @@ h1 {
 }
 
 main {
-  h2 {
+  h3 {
     line-height: 24px;
     padding-bottom: 20px;
     border-bottom: 1px solid lightgray;
